search populated local maven repo first if exists
If 'maven install' was run (from the console or within IDE), it's likely that all needed jars to run the example will be found below "$HOME/.m2".   Modify the bash snippet so that it maybe used instead, if the install phase has run.  If it has not (the jars have been built to target/ subdirectories only), rely on the previous locations to search.