fix face culling when fence and wall drawing, fix fence texture if available width is <1
this draws two textured faces for each fence "side" (spaced
zero width currently), but removes / culls the faces seen from
the back side for each frame viewed.  since the faces both lie
in the same plane but fronts are oriented opposite, exactly one
face is viewed, while the other is removed (applies to each
possible view).

face culling the faces currently viewed from the back side
is less performant than not having had them added to the
geometry in the first place (prior art).

if face culling does not work (driver issue or other problem),
this is problematic, because the faces will be drawn at the same
position and it is random/undefined which pixel is viewed.  this
is noticed as dirty flickering if the viewport is moved.

in such a non-working state, you are better off with prior art,
the previous method, that draws one face only which is then viewed
from both sides (back and front, without culling), experiment with
- model.useLight = false / true
- or flipping the normals depending on the light position
  (can be checked by calculating the dot product of given
   normal and vector to light source, if it is <0, normal
   should be flipped)
  (you will also need two-sided lighting in this case, or
   the computed light intensity will be the same on both
   sides of the fence at all times)