make clipping adjustable, don't eat cpu time if there is nothing new to draw
Using the mouse wheel clipping is now adjustable, in addition
zooming in and out was added (adjustable field of width).

- CTRL + mouse wheel changes the zFar clipping, this is important when viewing
  the world from great height, changes are made persistent through JOSM prefs
- CTRL + SHIFT + mouse wheel uses greater increment (change happens faster)

- mouse wheel movement changes the zoom by using field of view adjustments
- SHIFT + mouse wheel changes the zoom quicker

Previously a lot of cpu cycles were wasted by running the animator when there
was nothing new to draw, which especially lead to high cpu loads if a lot of
objects were to be drawn.  This can decrease performance if the cpu heats up
and lowers its clock frequency in order to protect from overheating.  In a
normal usage pattern with input pauses, the cpu now gets time to cool down
while kendzi3d is used.

This commit tames the GL render loop by pausing the animator if the last
camera update did not result in a camera state change.

There is still work to do queuing correct resume events when buildings
are selected in the viewport (i.e. the "selection listener"), but camera
movement and rotations are correctly handled, including kinematics.

Careful:  If you work on this, keep in mind that e.g. reshape events are
not consumed / processed while the GL animator is paused.  This is why
an AWT ComponentListener handles resize events to the GLCanvas window,
this listener resumes the GL animator on demand.