defer animator resuming until DataEvent is processed
- refactor Kendzi3dAction to Resumer
- let DataEvent implement Resumer

Previously GLCanvas animator was resumed when DataEvents were queued.
Now the DataEvent can be used to resume the canvas which makes it
possible to defer the resumal until after the event has been process()ed
by EditorObjectsProducer.  SelectionSynchronizeManager was changed to
implement DataEventListener, but the add() method continues to delegate
to selectionChanged(), this makes Kendzi3dGlFrame.add() DataEvent
handling routing more readable.

-----
The process() and combine() methods of EditorObjectsProducer.java
do not differentiate between NewDataEvent and UpdateDataEvent to its
full potential, yet.

* filteredIds = DataSetFilterUtil.filter(layer, dataSet, perspective) is
always run on the full edit dataset of the active edit layer, i. e.
filteredIds is calculated in the same way for both event types.  If
NewDataEvent is process()ed, the data layers (previous editableObjects)
are cleared before combine(), this is the only major difference in
editor object creation / updating.

* filteredIds = DataSetFilterUtil.filter(layer,
getJosmEvent().getAffectedPrimtives(), perspective) is desirable to
have, if an UpdateDataEvent is received however, i.e. to use the
primitives attached to the AbstractDataChangedEvent received.

The set of affected primitives in the UpdateDataEvent case is much
smaller than the whole active edit layer dataset. UpdateDataEvents
are produced numerously. Running DataSetFilterUtil.filter() on the
affected primitives only, rather than the whole dataset, is likely
to boost performance of the render loop.

-----
Not using the whole dataSet, the code has to resolve primitive
dependencies in some cases, e.g. NodeMovedEvent.getPrimitives() has to
be post-processed, as long as DataSetFilterUtil.filter() is unchanged.

* Looking solely at the Node filter() will miss geometry changes of
objects that are tracked using ids of higher level way or relation
primitives, that refer to Node.

* Instead of enriching the primitives collection input to filter(), by
using getReferrers() on Node multiple times, an alternative is to track
HashMap<Node, List<WorldObject>> or similar, which may offer more
performance at the cost of memory and the additional burden of tracking
changes to the HashMap.

-----
Set aside the HashMap approach, work-in-progress on improved
UpdateDataEvent handling has already been done. To better test its
behavior and give the possibility to compare or
measure performance implications,

* unified event handling will remain untouched within this repo branch

* reworked UpdateDataEvent handling code in another branch