fix a map center update bug, check if editDataSet changed
* if the editDataSet was changed, process a NewDataEvent in any case
* fix move camera action, if perspective center remained at its initial
  LatLon.ZERO after data was added, it produced display artifacts:

If a new data layer is created in JOSM without data, the center used is
LatLon.ZERO.  This is true, if the user opens the 3d window, before any
data was added to the data layer.

If osm primitives are added to the data layer, then the UpdateDataEvents
process()ed by the 3d plugin in response, do not lead to recalculation
of the center variable or perspective.

After adding e.g. a road, the usual workflow is to use "move camera"
from the 3d menu in JOSM.  If the camera is moved based on a perspective
calculated with LatLon.ZERO to a distant position, the display shows
wobbly artifacts, a jumping compass in the lower corner and huge texture
pixels are other noticable symptoms.  Closing the window and reopening
does not help either, because the perspective is only recalculated in
NewDataEvent case.

With this commit, when events are processed in EditorObjectsProducer, as
long as the center is stuck at LatLon.ZERO, any event is treated as a
NewDataEvent.