remove the interim dep on josm from kendzi3d-jogl-util
The dependency was added to lookup and store values in josm prefs, but
kendzi3d-jogl-util should be compilable and usable without a josm
dependency.

For this reason CameraMoveListener and Viewport were subclassed and the
subclasses are now stored within the scope of kendzi3d-plugin java
project.  All functionality that needed to read/write josm prefs was
moved to the subclasses.

The pom.xml and the classes within kendzi3d-jogl-util remain independent
of a josm dependency.  For example, kendzi3d-jogl-example depends also
on kendzi3d-jogl-util, but not on josm.  The previous changes made the
jogl-example indirectly depend on josm classes which was a negative side
effect that this commit fixes.