subclassing Viewport suffices, use unified CameraMoveListener class
The commit documents how to run the example editor (kendzi3d-editor)
standalone, outside of an IDE like eclipse. It uses a Viewport class
that does not depend on JOSM for pref storage, like Kendzi3dViewport
does in contrast.

Since the CameraMoveListener is generic to all sub-projects (again),
zooming and adjusting the distance to the far clipping plane is now
available in the example editor as well.  (But only kendzi3d-plugin
project saves viewport related settings across restarts, employing
josm prefs.)

The simple example (kendzi3d-jogl-example) does not make use of the
Viewport class, it does use CameraMoveListener with the viewport set
to null. Hence it does not exhibit functional changes.